Added the new database schema with update/refresh system.
The new schema is to be used in conjunction with R to plot graphs. The new
tables include network_size, relay_versions, relay_platforms, and
total_bandwidth, which can be queried by R so complex joins can be avoided
for the sake of speed. The new system keeps these summary tables up-to-date
by keeping track of when changes were made in the descriptor or statusentry
tables.

Moved everything database related to db/

Modified run.sh to use refresch script when ernie runs.

Add Fast and Stable flag to network_size table.

Add versions 0.1.0, 0.1.1, and 0.2.3 to relay_versions.