Migrating materialized views to calculate deltas from last update.
Instead of a true materialized view system where we select all of the
date from the views every time, it will be faster to select just the
most recent updates to the table, and then append them to the summary
tables used for the time graphs.

Removed materialized views from schema and moved refresh functions to tordir.

The materialized views are being superceded by a system that updates
the tables based on the differences in the newest entries because
it is faster.