Clean up the auto-generated tordir.sql a bit.
Remove PostgreSQL's settings which initially came from the schema export.
We should only change a default if there's a good reason to so. If we
find such a reason, we should put them back and add a comment why we think
we need them.

Put SQL keywords in all caps and remove auto-generated comments.

Remove auto-generated GRANT and REVOKE statements as well as indexes which
are generated from the primary key constraints anyway.