Tweak Kevin's new database schema a bit.
Note: This commit can and should be squashed together with Kevin's
commits!

Changes:

Update the relay_statuses_per_day table with a refresh_*() function, too.

Add a note to tordir-denormalized.sql that we're currently not using it.

Add Fast and Stable flag to network_size table.

Add versions 0.1.0, 0.1.1, and 0.2.3 to relay_versions.

Take out platforms_uptime_month and the corresponding refresh function
which should be reviewed and merged together with the churn stuff.
BE SURE TO BACKUP THESE BEFORE SQUASHING!!

Fix update_status() to work for UPDATE, too.

Add publication date of descriptors plus the subsequent day to updates,
too.

Exclude all relays without the Running flag from all statistics by
requiring that Running is set in refresh_*() functions.

Change 'descriptor LEFT JOIN statusentry' to RIGHT JOIN, because we don't
want to consider unreferenced descriptors, but we do want to consider
status entries even when we are missing a descriptor.